split asset icon fetching from presentational component
