+++ src/pages/PoolsOverview.tsx
@@ -1,19 +1,14 @@
 import { Layout } from "@stellar/design-system";
-import { Avatar } from "components/Avatar";
+import { AssetAvatar } from "components/AssetAvatar";
 
 export const PoolsOverview = () => (
   <Layout.Inset>
     Pools overview page
     <div style={{ marginTop: 20 }}>
-      <Avatar
-        source={[
-          {
-            asset:
-              "USDC:GA5ZSEJYB37JRC5AVCIA5MOP4RHTM335X2KGX3IHOJAPP5RE34K4KZVN",
-          },
-          {
-            asset: "native",
-          },
+      <AssetAvatar
+        assets={[
+          "USDC:GA5ZSEJYB37JRC5AVCIA5MOP4RHTM335X2KGX3IHOJAPP5RE34K4KZVN",
+          "native",
         ]}
       />
     </div>
